Refactor ClusterInformation
